+++ pkg/accountmanager/accountmanageriface/mocks/Servicer.go
@@ -12,6 +12,20 @@ type Servicer struct {
 	mock.Mock
 }
 
+// DeletePrincipalAccess provides a mock function with given fields: _a0
+func (_m *Servicer) DeletePrincipalAccess(_a0 *account.Account) error {
+	ret := _m.Called(_a0)
+
+	var r0 error
+	if rf, ok := ret.Get(0).(func(*account.Account) error); ok {
+		r0 = rf(_a0)
+	} else {
+		r0 = ret.Error(0)
+	}
+
+	return r0
+}
+
 // UpsertPrincipalAccess provides a mock function with given fields: _a0
 func (_m *Servicer) UpsertPrincipalAccess(_a0 *account.Account) error {
 	ret := _m.Called(_a0)