+++ pkg/common/commoniface/codebuild.go
@@ -1,7 +1,0 @@
-package commoniface
-
-import "github.com/aws/aws-sdk-go/service/codebuild/codebuildiface"
-
-type CodeBuildAPI interface {
-	codebuildiface.CodeBuildAPI
-}